requirements.txt
# Function dependencies, for example:
# package>=version
google-api-python-client==1.7.8
google-auth-httplib2==0.0.3
google-auth==1.6.2
oauth2client==4.1.3